Refactor out some custom type for cli
