+++ src/commands/models/mod.rs
@@ -1,0 +1,5 @@
+pub mod directory;
+pub mod root_directory;
+
+pub use directory::*;
+pub use root_directory::*;
