+++ src/path.rs
@@ -1,8 +1,6 @@
 use super::config::Config;
-use std::fmt;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::str::FromStr;
 
 fn config_dir() -> Option<PathBuf> {
     let config_dir = dirs::config_dir().map(|p| p.join("dadmin"))?;
@@ -22,32 +20,6 @@ pub fn user_path() -> Option<PathBuf> {
     Some(config)
 }
 
-pub fn validate_root(root: &str) -> Result<String, RootError> {
-    let path = Path::new(root);
-
-    if path.is_relative() {
-        return Err(RootError::NotAnAbsolute {
-            path: root.to_string(),
-        });
-    }
-    if path.exists() {
-        if path.is_dir() {
-            Ok(root.to_string())
-        } else {
-            Err(RootError::RootIsAFile {
-                path: root.to_string(),
-            })
-        }
-    } else {
-        fs::create_dir_all(root)
-            .map(|_| root.to_string())
-            .map_err(|source| RootError::CannotCreateRoot {
-                source,
-                path: root.to_string(),
-            })
-    }
-}
-
 pub fn local_path(organisation: &str, name: &str) -> Option<PathBuf> {
     let root = Config::root().ok()?;
     let root_dir = Path::new(&root);
@@ -62,51 +34,6 @@ pub fn local_path_org(organisation: &str) -> anyhow::Result<PathBuf> {
     Ok(local_path)
 }
 
-#[derive(thiserror::Error, Debug)]
-pub enum RootError {
-    #[error("{path} is a file. Root directory cannot be a file")]
-    RootIsAFile { path: String },
-    #[error("Cannot create root directory for path: {path} with error: {source}")]
-    CannotCreateRoot {
-        source: std::io::Error,
-        path: String,
-    },
-    #[error("{path} is not an absolute path. Root must be an absolute path")]
-    NotAnAbsolute { path: String },
-}
-
-#[derive(Debug)]
-pub struct RootDirectory {
-    pub path: String,
-}
-
-impl FromStr for RootDirectory {
-    type Err = RootError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        validate_root(s).map(|path| RootDirectory { path })
-    }
-}
-
-impl Default for RootDirectory {
-    fn default() -> Self {
-        RootDirectory {
-            path: dirs::home_dir()
-                .expect("Cannot have unknown home directory")
-                .join("dadmin")
-                .to_str()
-                .expect("Non UTF8 path is not supported right now")
-                .to_string(),
-        }
-    }
-}
-
-impl fmt::Display for RootDirectory {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.path)
-    }
-}
-
 trait EnsureDirExists: Sized {
     fn ensure_dir_exists(self) -> std::io::Result<Self>;
 }
@@ -118,40 +45,3 @@ impl EnsureDirExists for std::path::PathBuf {
     }
 }
 
-#[derive(Debug)]
-pub struct Directory {
-    pub path: PathBuf,
-}
-
-impl FromStr for Directory {
-    type Err = DirError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        validate_dir(s).map(|path| Directory { path })
-    }
-}
-#[derive(thiserror::Error, Debug)]
-pub enum DirError {
-    #[error("{path} is not a dir")]
-    NotADir { path: String },
-    #[error("{path} is not exist")]
-    NotExist { path: String },
-}
-
-pub fn validate_dir(dir: &str) -> Result<PathBuf, DirError> {
-    let path = Path::new(dir);
-
-    if path.exists() {
-        if path.is_dir() {
-            Ok(path.to_path_buf())
-        } else {
-            Err(DirError::NotADir {
-                path: dir.to_string(),
-            })
-        }
-    } else {
-        Err(DirError::NotExist {
-            path: dir.to_string(),
-        })
-    }
-}
